Add mac os support for CookieManager.cs

+++ Froststrap/Models/APIs/Roblox/AuthenticatedUser.cs
@@ -3,7 +3,7 @@
     public class AuthenticatedUser
     {
         [JsonPropertyName("id")]
-        public int Id { get; set; } = 0;
+        public long Id { get; set; } = 0;
 
         [JsonPropertyName("name")]
         public string Username { get; set; } = string.Empty;
